cloud_functions/video_handler/requirements.txt
openai-whisper
google-cloud-storage
google-cloud-pubsub
pytube3
functions-framework==3.*